+++ packages/runtime/src/index.ts
@@ -1,6 +1,7 @@
 export * from "./runtime-dsl-parser";
 export * from "./compiler/view-compiler";
 export * from "./compiler/plan-builder";
+export * from "./dsl/expression";
 export * from "./graph/dependency-graph";
 export * from "./graph/topo-sort";
 export * from "./graph/dep-resolver";